Opportunistically move OfflineBillingController to the standard location as well
